requirements.txt
# Core dependencies
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
loguru>=0.7.0
click>=8.1.0
dependency-injector>=4.41.0

# File parsing
openpyxl>=3.1.0
xlrd>=2.0.1

# Data processing
numpy>=1.24.0
python-dateutil>=2.8.2

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0 